safari-600.7-branch/LayoutTests/media/track/track-webvtt-tc027-empty-cue-expected.txt

Tests that empty cues are discarded.


EXPECTED (video.textTracks[0].cues.length == '0'), OBSERVED '3' FAIL
END OF TEST
